configure routing on linux
