+++ inst/openvpn/misc_windows.go
@@ -23,3 +23,7 @@ func setRegValue(guid, name string) error {
 
 	return key.SetStringValue("Name", name)
 }
+
+func createNatRules(p, server string, port int) error {
+	return nil
+}
